requirements.txt
aioquic==1.2.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
backoff==2.2.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dynaconf==3.2.7
Faker==36.1.1
Flask==3.1.0
fqdn==1.5.1
graphql-core==3.2.6
h11==0.14.0
h2==4.1.0
harfile==0.3.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
hypothesis==6.130.5
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.10
iniconfig==2.0.0
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junit-xml==1.9
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==11.1.3
mitmproxy-macos==0.11.5
mitmproxy_rs==0.11.5
msgpack==1.1.0
multidict==6.2.0
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.3.1
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pylsqpack==0.3.19
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
pyrate-limiter==3.7.0
PySocks==1.7.1
pytest==8.3.4
pytest-subtests==0.14.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruff==0.11.2
schemathesis==3.39.14
selenium==4.29.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
starlette-testclient==0.4.1
tomli==2.2.1
tomli_w==1.2.0
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
urwid==2.6.16
wcwidth==0.2.13
webcolors==24.11.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
